+++ src/components/Input/index.ts
@@ -1,2 +1,3 @@
 export * from './TextField';
 export * from './NumberField';
+export * from './TextArea';